Moved test to unit/test_plugins.py
